requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-js-asset==3.1.2
django-tinymce==4.1.0
djangorestframework==3.15.2
mysql-connector-python==9.1.0
pillow==11.0.0
python-dotenv==1.0.1
sqlparse==0.5.2
